Refactor of testHelper and addition of first selection for dropdown type selection
